fix trpc caching
