+++ apps/client/src/pages/avatar/[id]/settings.tsx
@@ -16,13 +16,18 @@ export const getServerSideProps = async ({ res, query }: GetServerSidePropsConte
 
   res.setHeader(
     "Cache-Control",
-    `public, s-maxage=${ONE_MINUTE_IN_SECONDS}, stale-while-revalidate=${ONE_WEEK_IN_SECONDS}`
+    `public, max-age=0, s-maxage=${ONE_MINUTE_IN_SECONDS}, stale-while-revalidate=${ONE_WEEK_IN_SECONDS}`
   );
 
   const id = query.id as string;
-  const props = await getPublicationProps(id);
+  const publicationProps = await getPublicationProps(id);
 
-  return { props };
+  return {
+    props: {
+      id,
+      ...publicationProps,
+    },
+  };
 };
 
 export default function Settings(props: InferGetServerSidePropsType<typeof getServerSideProps>) {
@@ -39,6 +44,11 @@ export default function Settings(props: InferGetServerSidePropsType<typeof getSe
       refetchOnWindowFocus: false,
       refetchOnMount: false,
       refetchOnReconnect: false,
+      trpc: {
+        context: {
+          skipBatch: true,
+        },
+      },
     }
   );
